test: split the unreproducible unit tests along the job split
`manual_deb_sync`'s detection, the shared `UnreproducibleSyncJob` half and
`manual_installs_sync`'s unowned scan each get their own module, with the fixtures
all three share in `unreproducible_harness.py`.
